Refactor: centralize JSON helpers
