+++ plant_engine/utils.py
@@ -1,0 +1,18 @@
+import json
+import os
+from typing import Any, Dict
+
+__all__ = ["load_json", "save_json"]
+
+
+def load_json(path: str) -> Dict[str, Any]:
+    """Load a JSON file and return its contents."""
+    with open(path, "r", encoding="utf-8") as f:
+        return json.load(f)
+
+
+def save_json(path: str, data: Dict[str, Any]) -> None:
+    """Write a dictionary to a JSON file, creating parent dirs if needed."""
+    os.makedirs(os.path.dirname(path), exist_ok=True)
+    with open(path, "w", encoding="utf-8") as f:
+        json.dump(data, f, indent=2)
