+++ scripts/review_thresholds.py
@@ -1,6 +1,7 @@
 import os
 import json
-from approval_queue import apply_approved_thresholds, load_json, save_json
+from approval_queue import apply_approved_thresholds
+from plant_engine.utils import load_json, save_json
 
 PENDING_DIR = "data/pending_thresholds"
 PLANT_DIR = "plants"
